requirements.txt
alembic==1.8.0
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.0
click==8.1.3
distlib==0.3.4
dnspython==2.2.1
email-validator==1.1.3
filelock==3.7.1
Flask==2.0.1
Flask-Bcrypt==0.7.1
flask-expects-json==1.7.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Seeder==1.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.0.3
jsonschema==4.6.0
Mako==1.2.0
MarkupSafe==2.1.1
pipenv==2022.6.7
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.1
python-decouple==3.4
six==1.16.0
SQLAlchemy==1.4.23
virtualenv==20.14.1
virtualenv-clone==0.5.7
Werkzeug==2.0.0
WTForms==3.0.1
